fix: added a validator and exception class for service, a missing method and tests
The query works correctly, I just refactored it a bit. I added a new method which retrieves artifacts by a category (I suppose we will modify the type of category from string to enum once we know exactly the categories we'll work with later). I also tested every method in the service.