feat: finished Section Controller, fixed working with LAZY
added a few queries and/or methods in `SectionRepository` such that LAZY loading works and does not throw `LazyInitializationException`; also changed `SectionService` accordingly.
the above mentioned are still subject to change, if somehow better ways to do things are found.

finished working on the `SectionController` endpoints; for now, they seem alright; still are subject to change; we have lines that are commented which have to do with authorization: e.g, only ARH users should be able to access a certain endpoint; can be uncommented later, when we'd finish testing.