feat: finished working on the Artifact controller
Apparently, the query in the Artifact repository was not retrieving the artifacts correctly, but I modified it, and it works properly now.
Added endpoints for every method in the service, for now there are no permissions because I wanted to test if they work as intended. Also, some of the endpoints were not mentioned in the issues, so I'm not entirely sure if their name are alright, I will modify them if needed.